Move StreamStateRegnerator out to the dataproviders module
